Moved header file, added Create Shift page, and created header with back button
