+++ src/App.js
@@ -18,6 +18,7 @@ import PastShifts from './components/pages/PastShifts';
 import Contacts from './components/pages/Contacts';
 import History from './components/pages/History';
 import PrivateRoute from './components/authentication/PrivateRoute';
+import CreateShift from './components/pages/CreateShift';
 
 // Your web app's Firebase configuration
 // For Firebase JS SDK v7.20.0 and later, measurementId is optional
@@ -69,6 +70,9 @@ export default function App() {
         <Route path="/success">
           <SuccessPage />
         </Route>
+        <PrivateRoute isLoggedIn={isLoggedIn} path="/schedule/create-shift">
+          <CreateShift />
+        </PrivateRoute>
         <PrivateRoute isLoggedIn={isLoggedIn} path="/schedule">
           <Schedule {...{ isAdmin }} />
         </PrivateRoute>
