+++ src/components/navigation/back-header.module.css
@@ -1,0 +1,7 @@
+.backArrow {
+  position: fixed;
+  width: 40px;
+  height: 40px;
+  left: 30px;
+  top: 32px;
+}
