+++ src/components/navigation/styled-heading.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const StyledHeading = styled.h2`
+  text-align: center;
+  padding-top: 32px;
+`;
+
+export default StyledHeading;
