+++ src/components/pages/Contacts.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import HeaderWithNav from './header';
+import HeaderWithNav from '../navigation/nav-header';
 
 export default function Contacts(props) {
   const { isAdmin } = props;
