+++ src/components/pages/Schedule.jsx
@@ -1,12 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import HeaderWithNav from './header';
+import { Link } from 'react-router-dom';
+import HeaderWithNav from '../navigation/nav-header';
 
 export default function Schedule(props) {
   const { isAdmin } = props;
   return (
     <div>
       <HeaderWithNav {...{ isAdmin }}>Schedule</HeaderWithNav>
+      <button type="button">
+        <Link to="schedule/create-shift">Create Shift</Link>
+      </button>
     </div>
   );
 }
